Final Report/Team10/NestWise-main/backend-RAG/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-core
langchain-community
langchain-openai
langgraph
tiktoken
numpy
scikit-learn
pypdf
unstructured
pdfminer.six
pymongo
python-jose[cryptography]
argon2-cffi
pydantic[email]
dnspython
python-jose